remove incorrect version suffix from brew install

+++ graph/model/models_gen.go
@@ -41,7 +41,7 @@ type FetchResult struct {
 // Returns the total number of commits in the repo and the array of 10 commit log entries
 type GitCommitLogResults struct {
 	// Total number of commits tracked by the branch
-	TotalCommits *float64 `json:"totalCommits"`
+	TotalCommits *int `json:"totalCommits"`
 	// Array of commit log entries `gitCommits`
 	Commits []*GitCommits `json:"commits"`
 }
@@ -70,7 +70,7 @@ type GitRepoStatusResults struct {
 	// Remote host name based on the remote URL (E.g: https://github.com/github/repo.git -> Github)
 	GitRemoteHost *string `json:"gitRemoteHost"`
 	// Total number of commits tracked by the current branch
-	GitTotalCommits *float64 `json:"gitTotalCommits"`
+	GitTotalCommits *int `json:"gitTotalCommits"`
 	// The latest commit (HEAD commit)
 	GitLatestCommit *string `json:"gitLatestCommit"`
 	// Total number of files tracked by git repo
